Update first-steps.md
It seems like we should briefly explain what the Bridge is here --I'm not sure that my definition is accurate, though.  Then we should explain what a sequence is before we tell them to click on it.  Also, Keptn needs to be capitalized and can't have apostrophe-s attached to it.